Removed display module
