+++ src/config/mod.rs
@@ -3,7 +3,6 @@
 //! CDDNS takes the typical layered configuration approach. There are 3 layers.
 //! The config file is the base, which is then superseded by environment
 //! variables, which are finally superseded by CLI arguments and options.
-mod display;
 pub mod models;
 
 /// Return the default configuration path, depending on the host OS. This may
